requirements.txt
streamlit==1.31.0
python-dotenv==1.0.0
boto3==1.34.34
Pillow==10.2.0
requests==2.31.0
python-jose==3.3.0
python-multipart==0.0.6
httpx==0.27.0
google-generativeai>=0.7.2
pytest==7.4.3
pytest-cov==4.1.0
moto==4.2.5
pytest-mock==3.12.0